+++ src/cofre/compartilhado/Resposta.java
@@ -16,9 +16,9 @@ public class Resposta implements Serializable {
 
     private final String  mensagem;
     private final boolean acerto;
-    private final int     premioOuFundo;
+    private final double     premioOuFundo;
 
-    public Resposta(String mensagem, boolean acerto, int premioOuFundo) {
+    public Resposta(String mensagem, boolean acerto, double premioOuFundo) {
         this.mensagem      = mensagem;
         this.acerto        = acerto;
         this.premioOuFundo = premioOuFundo;
@@ -26,5 +26,5 @@ public class Resposta implements Serializable {
 
     public String  getMensagem()      { return mensagem;      }
     public boolean isAcerto()         { return acerto;        }
-    public int     getPremioOuFundo() { return premioOuFundo; }
+    public double     getPremioOuFundo() { return premioOuFundo; }
 }
